function/requirements.txt
azure-functions
joblib==0.15.1
pandas==1.0.4
scikit-learn==0.22